move hammerspoon config
